+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 11.41
- - 
+ - Added a function to simplify hooking language strings from plugins
 
 ### 11.40
  - Code refactor of the heredocs used for the Evil Twin, WEP and WPS attacks